Refactor forum mocks
